requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic-settings>=2.4
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
redis>=5.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.9
itsdangerous>=2.1
email-validator>=2.0

# tests
pytest>=8.0
pytest-asyncio>=0.23
httpx>=0.27
aiosqlite>=0.20
fakeredis>=2.23
